packages/kbn-router-to-openapispec/src/util.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the "Elastic License
 * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
 * Public License v 1"; you may not use this file except in compliance with, at
 * your election, the "Elastic License 2.0", the "GNU Affero General Public
 * License v3.0 only", or the "Server Side Public License, v 1".
 */

// eslint-disable-next-line import/no-extraneous-dependencies
import { OpenAPIV3 } from 'openapi-types';
import {
  getRequestValidation,
  type RouteMethod,
  type RouteConfigOptions,
  type RouteConfigOptionsBody,
  type RouterRoute,
  type RouteValidatorConfig,
} from '@kbn/core-http-server';
import { CustomOperationObject, KnownParameters } from './type';
import type { GenerateOpenApiDocumentOptionsFilters } from './generate_oas';

const tagPrefix = 'oas-tag:';
const extractTag = (tag: string) => {
  if (tag.startsWith(tagPrefix)) {
    return tag.slice(tagPrefix.length);
  }
};
/**
 * Given an array of tags ([oas-tag:beep, oas-tag:boop]) will return a new array
 * with the tag prefix removed.
 */
export const extractTags = (tags?: readonly string[]) => {
  if (!tags) return [];
  return tags.flatMap((tag) => {
    const value = extractTag(tag);
    if (value) {
      return value;
    }
    return [];
  });
};

/**
 * Build the top-level tags entry based on the paths we extracted. We could
 * handle this while we are iterating over the routes, but this approach allows
 * us to keep this as a global document concern at the expense of some extra
 * processing.
 */
export const buildGlobalTags = (paths: OpenAPIV3.PathsObject, additionalTags: string[] = []) => {
  const tags = new Set<string>(additionalTags);
  for (const path of Object.values(paths)) {
    for (const method of Object.values(OpenAPIV3.HttpMethods)) {
      if (path?.[method]?.tags) {
        path[method]!.tags!.forEach((tag) => tags.add(tag));
      }
    }
  }
  return Array.from(tags)
    .sort((a, b) => a.localeCompare(b))
    .map<OpenAPIV3.TagObject>((name) => ({ name }));
};

export const getPathParameters = (path: string): KnownParameters => {
  return Array.from(path.matchAll(/\{([^{}?]+\??)\}/g)).reduce<KnownParameters>((acc, [_, key]) => {
    const optional = key.endsWith('?');
    acc[optional ? key.slice(0, key.length - 1) : key] = { optional };
    return acc;
  }, {});
};

export const extractContentType = (body: undefined | RouteConfigOptionsBody) => {
  if (body?.accepts) {
    return Array.isArray(body.accepts) ? body.accepts : [body.accepts];
  }
  return ['application/json'];
};

export const getVersionedContentTypeString = (
  version: string,
  acceptedContentTypes: string[]
): string => {
  return `${acceptedContentTypes.join('; ')}; Elastic-Api-Version=${version}`;
};

export const extractValidationSchemaFromRoute = (
  route: RouterRoute
): undefined | RouteValidatorConfig<unknown, unknown, unknown> => {
  if (!route.validationSchemas) return undefined;
  return getRequestValidation(route.validationSchemas);
};

export const getVersionedHeaderParam = (
  defaultVersion: undefined | string,
  versions: string[]
): OpenAPIV3.ParameterObject => ({
  in: 'header',
  name: 'elastic-api-version',
  description: 'The version of the API to use',
  schema: {
    type: 'string',
    enum: versions,
    default: defaultVersion,
  },
});

export const prepareRoutes = <
  R extends { path: string; options: { access?: 'public' | 'internal'; excludeFromOAS?: boolean } }
>(
  routes: R[],
  filters: GenerateOpenApiDocumentOptionsFilters = {}
): R[] => {
  if (Object.getOwnPropertyNames(filters).length === 0) return routes;
  return routes.filter((route) => {
    if (route.options.excludeFromOAS) return false;
    if (
      filters.excludePathsMatching &&
      filters.excludePathsMatching.some((ex) => route.path.startsWith(ex))
    ) {
      return false;
    }
    if (filters.pathStartsWith && !filters.pathStartsWith.some((p) => route.path.startsWith(p))) {
      return false;
    }
    if (filters.access && route.options.access !== filters.access) return false;
    return true;
  });
};

export const assignToPaths = (
  paths: OpenAPIV3.PathsObject,
  path: string,
  pathObject: OpenAPIV3.PathItemObject
): void => {
  const pathName = path.replace(/[\?\*]/, '');
  paths[pathName] = { ...paths[pathName], ...pathObject };
};

export const mergeResponseContent = (
  a: OpenAPIV3.ResponseObject['content'],
  b: OpenAPIV3.ResponseObject['content']
) => {
  const mergedContent = {
    ...(a ?? {}),
    ...(b ?? {}),
  };
  return { ...(Object.keys(mergedContent).length ? { content: mergedContent } : {}) };
};

export const getXsrfHeaderForMethod = (
  method: RouteMethod,
  options?: RouteConfigOptions<RouteMethod>
): OpenAPIV3.ParameterObject[] => {
  if (method === 'get' || method === 'options' || options?.xsrfRequired === false) return [];
  return [
    {
      description: 'A required header to protect against CSRF attacks',
      in: 'header',
      name: 'kbn-xsrf',
      required: true,
      schema: {
        example: 'true',
        type: 'string',
      },
    },
  ];
};

export const setXState = (
  availability: RouteConfigOptions<RouteMethod>['availability'],
  operation: CustomOperationObject
): void => {
  if (availability) {
    if (availability.stability === 'experimental') {
      operation['x-state'] = 'Technical Preview';
    }
    if (availability.stability === 'beta') {
      operation['x-state'] = 'Beta';
    }
  }
};

export type GetOpId = (input: { path: string; method: string }) => string;

/**
 * Best effort to generate operation IDs from route values
 */
export const createOpIdGenerator = (): GetOpId => {
  const idMap = new Map<string, number>();
  return function getOpId({ path, method }) {
    if (!method || !path) {
      throw new Error(
        `Must provide method and path, received: method: "${method}", path: "${path}"`
      );
    }

    path = path
      .trim()
      .replace(/^[\/]+/, '')
      .replace(/[\/]+$/, '')
      .toLowerCase();

    const removePrefixes = ['internal/api/', 'internal/', 'api/']; // longest to shortest
    for (const prefix of removePrefixes) {
      if (path.startsWith(prefix)) {
        path = path.substring(prefix.length);
        break;
      }
    }

    path = path
      .replace(/[\{\}\?\*]/g, '') // remove special chars
      .replace(/[\/_]/g, '-') // everything else to dashes
      .replace(/[-]+/g, '-'); // single dashes

    const opId = `${method.toLowerCase()}-${path}`;

    const cachedCount = idMap.get(opId) ?? 0;
    idMap.set(opId, cachedCount + 1);
    return cachedCount > 0 ? `${opId}-${cachedCount + 1}` : opId;
  };
};
